Merges functionality of TypedInsertion into Insertion and eliminates the former.  Replaces occurrences of "instanceof TypedInsertion" with getType() != null.
